requirements.txt
Django==1.10.1
Markdown==2.6.7
PyDrive==1.2.1
PyYAML==3.12
argparse==1.2.1
beautifulsoup4==4.5.1
bs4==0.0.1
dj-database-url==0.4.1
django-filter==0.15.2
djangorestframework==3.4.7
google-api-python-client==1.5.3
gunicorn==19.6.0
httplib2==0.9.2
oauth2client==3.0.0
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
whitenoise==3.2.1
wsgiref==0.1.2
